Be sure that a outputfilter post hook is always generated
Hi,
what about to change the behaviour of the hooks generation to always render it by minimum 1 to be sure that the hook get always rendered. In my opinion this makes the hooks much greater and i think that the current behaviour can make a lot of trouble, here my example:

You want to add code after the "DynamicField" block in different masks. If you now use the hook and some masks do not define dynamic fields in sysconfig then also the outputfilter will not match, because the hook does not exist.

Best regards,
Rolf